+++ packages/core/src/update-hooks.svelte.ts
@@ -26,7 +26,7 @@ export interface ContractUpdateOptions<S extends ContractSchemas> {
   overtimeOptions?: OvertimeOptions;
 }
 type Result<S extends ContractSchemas> = { data: ContractRecord<S> };
-interface Callbacks<S extends ContractSchemas> {
+export interface Callbacks<S extends ContractSchemas> {
   onSuccess?: (data: Result<S>, params: ContractUpdateParams<S>) => void | Promise<void>;
   onError?: (error: HttpError, params: ContractUpdateParams<S> | undefined) => void | Promise<void>;
   onSettled?: (data: Result<S> | undefined, error: HttpError | null, params: ContractUpdateParams<S> | undefined) => void | Promise<void>;